requirements.txt
streamlit
pandas
tensorflow
torch
transformers
scikit-learn
nltk
pillow
matplotlib
faiss-cpu
tf-keras
sentence-transformers